fix(database-settings): move to service-settings domain
