+++ libs/domains/service-settings/feature/src/index.ts
@@ -14,3 +14,4 @@ export * from './lib/helm-networking-settings/helm-networking-settings'
 export * from './lib/job-configuration/job-configuration'
 export * from './lib/terraform-variables-settings/terraform-variables-settings'
 export * from './lib/job-dockerfile-settings/job-dockerfile-settings'
+export * from './lib/database-settings-resources/database-settings-resources'
