+++ libs/shared/console-shared/src/index.ts
@@ -9,5 +9,4 @@ export * from './lib/flow-create-port/ui/crud-modal/crud-modal'
 export * from './lib/flow-create-port/utils/port-healthcheck'
 export * from './lib/layout-logs/layout-logs'
 export * from './lib/layout-logs/card-cluster/card-cluster'
-export * from './lib/database-settings-resources/ui/database-settings-resources/database-settings-resources'
 export * from './lib/settings-heading/settings-heading'
